requirements.txt
python-dotenv==0.20.0
pandas==1.4.3
fastapi==0.78.0
gunicorn==20.1.0
uvicorn==0.18.2